MAJOR update.
Handled multithreading for sending commands and variables to KMuddy.

Figuring out a plan for receiving variables in a multithreaded way,
and though I haven't found a need for it yet, someone else probably will :)

Added auto-walking to the room ratter, thus giving the ability
to fully automate ratting. (ILLEGAL!)'

Added comments to all of the plugins for easier understanding :)

Lots of other things.